refactor: move error in dedicated file
