+++ src/lib.rs
@@ -24,9 +24,11 @@
 pub extern crate bip39;
 pub extern crate bitcoin;
 
+pub mod error;
+
+use error::*;
 use std::convert::TryInto;
 use std::default::Default;
-use std::fmt;
 
 use bitcoin::bip32;
 use bitcoin::bip32::ChildNumber;
@@ -40,45 +42,6 @@ use bitcoin::PrivateKey;
 use bip39::Language;
 #[cfg(feature = "mnemonic")]
 use bip39::Mnemonic;
-
-/// A BIP85 error.
-#[derive(Clone, PartialEq, Eq)]
-pub enum Error {
-    /// Hardened index is provided, but only non-hardened indexes are allowed
-    InvalidIndex(u32),
-    /// Wrong number of bytes requested
-    InvalidLength(u32),
-    /// Wrong number of words for mnemonic
-    InvalidWordCount(u32),
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            Error::InvalidIndex(index) => write!(
-                f,
-                "invalid index for derivation, should be less than 0x80000000: {}",
-                index,
-            ),
-            Error::InvalidLength(len) => write!(
-                f,
-                "invalid bytes length: {}. Should be between 16 and 64",
-                len,
-            ),
-            Error::InvalidWordCount(word_count) => write!(
-                f,
-                "invalid number of words for mnemonic: {}. Should be 12, 18 or 24",
-                word_count,
-            ),
-        }
-    }
-}
-
-impl fmt::Debug for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        fmt::Display::fmt(self, f)
-    }
-}
 
 /// Derive raw bytes from the root inner using provided derivation path.
 ///
